Remove python module zstd dep in favour of libzef bundled version
