+++ python/requirements.txt
@@ -12,7 +12,6 @@ pytz>=2021.1
 more-itertools>=4.2.0
 art>=5.2
 PyYAML>=6.0
-zstd>=1.5.1.0
 pandas>=1.3.5
 case-converter>=1.0.2
 watchdog>=2.1.9
