+++ python/zef/core/internals/__init__.py
@@ -70,11 +70,13 @@ from ...pyzef.internals import (
     all_relation_types,
     apply_update,
     blob_to_json,
+    compress_zstd,
     create_graph_from_bytes,
     create_update_heads,
     create_update_payload,
     created_token_list,
     current_zefdb_protocol_version,
+    decompress_zstd,
     delete_graphdata,
     delegate_to_ezr,
     early_token_list,
